+++ panels/C1_Visualize/1_visualize-panel-ui.R
@@ -260,7 +260,24 @@ vis.mainPanel <- function() {
           column(10, hr())
         ),
         helpText("Statistical Summary for the data."),
-        verbatimTextOutput("visualize.summary")
+        verbatimTextOutput("visualize.summary"),
+        helpText("Formatting options"),
+        fixedRow(
+          column(
+            2,
+            numericInput("global.sig.level",
+              label = "Signifcant figures",
+              value = graphical.par$signif
+            )
+          ),
+          column(
+            2,
+            numericInput("global.round.pct",
+              label = "Round percentages",
+              value = graphical.par$round_percent
+            )
+          )
+        )
       ),
       ##  Inference Panel
       tabPanel(
@@ -278,8 +295,8 @@ vis.mainPanel <- function() {
         fixedRow(
           column(3, uiOutput("inference_type")),
           column(3, uiOutput("inference_test")),
-          column(2, uiOutput("inference_out")),
-          column(3, uiOutput("ci_width")),
+          column(6, uiOutput("inference_out")),
+          column(9, uiOutput("inference_opts")),
           column(3, uiOutput("inference_epi"))
         ),
         br(),